backend/pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "knowbase-backend"
version = "0.1.0"
description = "KnowBase AI Personal Knowledge Base – FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi[standard]>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pgvector>=0.3.2",
    "psycopg2-binary>=2.9.9",

    # Pydantic / Settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Auth / Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # AI / LLM
    "openai>=1.35.0",
    "langchain>=0.2.6",
    "langchain-openai>=0.1.14",
    "langchain-community>=0.2.6",
    "langgraph>=0.1.14",
    "tiktoken>=0.7.0",

    # Vector stores
    "qdrant-client>=1.9.1",

    # Document parsing
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.0",
    "html2text>=2024.2.26",
    "pillow>=10.3.0",
    "pytesseract>=0.3.10",
    "markdownify>=0.13.1",

    # Queue / worker
    "celery>=5.4.0",
    "redis>=5.0.6",
    "kombu>=5.3.7",

    # Utilities
    "python-slugify>=8.0.4",
    "tenacity>=8.4.1",
    "structlog>=24.2.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-passlib",
    "types-python-jose",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
